Removed Abscissa testing, cleaned up CI folder
