--- .github/workflows/integration.yaml
@@ -1,52 +1,0 @@
-name: Integration test
-on: [pull_request]
-jobs:
-  test-integration-stable:
-    runs-on: ubuntu-latest
-    services:
-#      chain_a:
-#        image: informaldev/chain_a
-#        ports:
-#          - 26656:26656
-#          - 26657:26657
-#      chain_b:
-#        image: informaldev/chain_b
-#        ports:
-#          - 26556:26656
-#          - 26557:26657
-      simd:
-        image: informaldev/simd
-        ports:
-          - 1317:1317
-          - 26656:26656
-          - 26657:26657
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-          override: true
-      - uses: actions-rs/cargo@v1
-        with:
-          command: build
-          args: -p relayer-cli --bins
-#      - name: Set up chain_a
-#        run: |
-#          TRUSTED_HEADER="$(curl -s http://localhost:26657/status)"
-#          HASH="$(echo "${TRUSTED_HEADER}" | jq -r .result.sync_info.latest_block_hash)"
-#          HEIGHT="$(echo "${TRUSTED_HEADER}" | jq -r .result.sync_info.latest_block_height)"
-#          cargo run --bin relayer-cli -- -c relayer/tests/config/fixtures/relayer_conf_example.toml light init -x "${HASH}" -h "${HEIGHT}" chain_A
-#      - name: Set up chain_b
-#        run: |
-#          TRUSTED_HEADER="$(curl -s http://localhost:26557/status)"
-#          HASH="$(echo "${TRUSTED_HEADER}" | jq -r .result.sync_info.latest_block_hash)"
-#          HEIGHT="$(echo "${TRUSTED_HEADER}" | jq -r .result.sync_info.latest_block_height)"
-#          cargo run --bin relayer-cli -- -c relayer/tests/config/fixtures/relayer_conf_example.toml light init -x "${HASH}" -h "${HEIGHT}" chain_B
-#      - name: Run relayer in the background
-#        run: |
-#          cargo run --bin relayer -- -c relayer/tests/config/fixtures/relayer_conf_example.toml start --reset &
-#          sleep 3
-      - uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: -p relayer-cli --test acceptance --no-fail-fast -- --ignored
